src/requirements.txt
aiosqlite==0.20.0
alembic==1.13.2
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
apprise==1.9.0
asgi-lifespan==2.1.0
async-timeout==4.0.3
asynch==0.2.4
asyncpg==0.29.0
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
ciso8601==2.3.1
click==8.1.7
clickhouse-connect==0.7.19
clickhouse-driver==0.2.9
clickhouse-sqlalchemy==0.3.2
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.3.0
coolname==2.2.0
croniter==3.0.3
cryptography==43.0.1
cycler==0.12.1
dateparser==1.2.0
docker==7.1.0
exceptiongroup==1.2.2
fastapi==0.114.0
Flask==3.0.3
Flask-Cors==5.0.0
fonttools==4.53.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
graphviz==0.20.3
greenlet==3.0.3
griffe==1.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
humanize==4.10.0
hyperframe==6.0.1
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jinja2-humanize-extension==0.4.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kafka
kafka-python>=2.0.2
kiwisolver==1.4.7
leb128==1.0.8
lz4==4.3.3
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
narwhals==1.6.2
numpy
oauthlib==3.2.2
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
pendulum==3.0.0
pika==1.3.2
pillow==10.4.0
plotly==5.24.0
pluggy==1.5.0
prefect==3.0.1
prometheus_client==0.20.0
protobuf==5.28.0
psycopg2-binary==2.9.9
py4j==0.10.9.7
pyarrow==17.0.0
pycparser==2.22
pydantic==2.9.1
pydantic-extra-types==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.23.3
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
pyspark==3.5.2
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.2
readchar==4.2.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==13.8.0
rpds-py==0.20.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn
scipy
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.5
streamlit==1.38.0
tenacity==8.5.0
text-unidecode==1.3
threadpoolctl==3.5.0
time-machine==2.15.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.6
watchdog==2.1.5
websockets==13.0.1
Werkzeug==3.0.4
zstandard==0.23.0
zstd==1.5.5.1
